Reset the audio stream volume to fix low in-call volume bug
Due to a bug in the OMX system, the audio stream volume is set to 0
after it was set to it's default volume.
Calling PhoneUtils.resetAudioStreamVolume() triggers the system to
reset the volume.

This should be called on every place where is switched between
audio modes.

REMARK: I think it only appears on the voice call stream,
        but to be sure I also do it on the bluetooth stream.

PS2:  Make low in-call volume fix configurable
      By making it configurable, it can be integrated in Rom source without
      impact on other devices which don't have the problem

Change-Id: Ieaee7d0dfde74ef815f969e01179b5a956c1adb7